requirements.txt
fastapi[all]
psycopg2-binary
sentence-transformers
pydantic
sqlalchemy
asyncpg
alembic
python-dotenv
greenlet